Buffer overflow in cursor create_name_map (#1)
Function create_name_map used a static buffer `szName[300]` for the column name in call to `SQLDescribeColW`. That function would truncate the column name to the size of the provided buffer, but return the actual length of the column name in the `&cchName` return parameter.
That `cchName` output was used to compute `cbName` parameter for the call to `TextBufferToObject(enc, szName, cbName)`. When the actual column name was longer than the buffer, cbName was to big, and was causing a buffer overflow, resulting in crashes, invalid data, or errors.

Fix this by making szName a dynamic buffer. This buffer overflow is very similar to the one fixed in https://github.com/mkleehammer/pyodbc/pull/881. I check through the rest of the code base, and I don't see more places with static buffers that could have a similar issue.

Upstream PR: https://github.com/mkleehammer/pyodbc/pull/931